func_geo v0.0.1 FuncGeo

Functional Geometry

This work is based on the Functional Geometry paper by Peter Henderson, the most recent version of this document is from 2002, the original version is from 1982. In this paper Peter Henderson deconstructs the M.C. Escher’s woodcut Square Limit using an algebra of pictures.

Peter Henderson also mentions that this idea can be applied to many more complex domains such as music or animation.

Other implementations

Lisp by Frank Buß

F# by Satnam Singh

Python and Haskell by Will McCutchen

Julia by Shashi Gowda

Functions

above(p, q)

Returns picture that has p in the upper half of its locating box and q in the lower half

above(m, n, p, q)

Place picture p above q scaled by m and n

as_ps(list)

as_svg(list)

beside(p, q)

Returns picture that has p in the left half of its locating box and q in the right half

beside(m, n, p, q)

Place picture p besides q scaled by m and n

blank()

Represents an empty picture

cycle(p)

Returns four times the p, layouted in a square and rotated

flip(p)

Returns picture p flipped through its vertical centre axis

grid(m, n, s)

Defines a picture function from lines in a grid

nonet(p, q, r, s, t, u, v, w, x)

over(p, q)

Overlays one picture p directly on top of the other picture q

plot(p, options \\ [])

Writes the given picture in the given format.

polygon(points)

Convert the given points, which specifies a polygon, in a list of lines

quartet(p, q, r, s)

Returns the picture p, q, r and s, layouted in a square

rot(p)

Returns picture p rotated anticlockwise by 90 degrees

rot45(p)

Rotates a picture about its top left corner, through 45 degrees anticlockwise. It also reduces the picture is size by sqrt(2)